website/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.6
Django==4.1.5
django-admin-list-filter-dropdown==1.0.3
django-autocomplete-light==3.9.7
django-tqdm==1.3.1
filelock==3.9.0
fonttools==4.55.3
fsspec==2024.12.0
huggingface-hub==0.27.1
idna==3.4
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pandas==2.2.3
Pillow==10.0.0
pipenv==2023.4.20
platformdirs==2.6.2
psycopg2-binary==2.9.9
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.31.0
safetensors==0.5.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
sympy==1.13.1
timm==1.0.13
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2022.7
urllib3==2.0.4
virtualenv==20.17.1
virtualenv-clone==0.5.7
whitenoise==6.8.2
Wikipedia-API==0.6.0
